フォルダ整理とスライダー完成
