+++ Assets/Graphics/α_version/Animation/Pitcher.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8c1b3d70e5a0e67429f7a8000dab89f4
+guid: 05c8c0b08d39c70448e1b33b96cbe834
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
