+++ Assets/Graphics/α_version/Animation/Pitcher/PitcherIdle.anim
