+++ Assets/Graphics/α_version/Animation/Pitcher/Throw.anim.meta
