+++ Assets/Graphics/α_version/Materials/Pitcher.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ecb9ffb805b2ee34ebdb520676ed2fdd
+guid: 2752240d57679ed41ac027c1e0133f4e
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
