+++ Assets/Graphics/α_version/Materials/Pitcher/Beta_Joints_MAT.mat
