+++ Assets/Graphics/α_version/Model/Pitcher.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c0bbca1e81eae0c4d8d560173fe61175
+guid: e2c66b7c66de2464fa250be78831c751
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
