+++ Assets/Graphics/α_version/Model/Pitcher/xbot@Goalie Throw.fbx
