--- Assets/Junjun/Scenes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cf99d4f5bf2f6fa479daefa78770e193
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
