--- Assets/Junjun/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 51699a629fc958142a6a0b5439db868a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
